refactor(vitest): move public exports to public folder
