--- packages/vitest/src/coverage.ts
@@ -1,1 +1,0 @@
-export { BaseCoverageProvider } from './utils/coverage'
