+++ packages/vitest/src/public/browser.ts
@@ -3,11 +3,11 @@ export {
   setupCommonEnv,
   loadDiffConfig,
   loadSnapshotSerializers,
-} from './runtime/setup-common'
+} from '../runtime/setup-common'
 export {
   takeCoverageInsideWorker,
   stopCoverageInsideWorker,
   getCoverageProvider,
   startCoverageInsideWorker,
-} from './integrations/coverage'
-export * as SpyModule from './integrations/spy'
+} from '../integrations/coverage'
+export * as SpyModule from '../integrations/spy'
