+++ packages/vitest/src/public/coverage.ts
@@ -1,0 +1,1 @@
+export { BaseCoverageProvider } from '../utils/coverage'
