+++ packages/vitest/src/public/runners.ts
@@ -1,0 +1,2 @@
+export { VitestTestRunner } from '../runtime/runners/test'
+export { NodeBenchmarkRunner } from '../runtime/runners/benchmark'
