+++ packages/vitest/src/public/snapshot.ts
@@ -1,0 +1,2 @@
+export type { SnapshotEnvironment } from '@vitest/snapshot/environment'
+export { VitestNodeSnapshotEnvironment as VitestSnapshotEnvironment } from '../integrations/snapshot/environments/node'
