+++ packages/vitest/src/public/workers.ts
@@ -1,0 +1,10 @@
+export {
+  createForksRpcOptions,
+  createThreadsRpcOptions,
+  unwrapSerializableConfig,
+} from '../runtime/workers/utils'
+export { provideWorkerState } from '../runtime/utils'
+export { run as runVitestWorker, collect as collectVitestWorkerTests } from '../runtime/worker'
+export { runVmTests } from '../runtime/workers/vm'
+export { runBaseTests } from '../runtime/workers/base'
+export type { WorkerRpcOptions, VitestWorker } from '../runtime/workers/types'
