--- packages/vitest/src/runners.ts
@@ -1,2 +1,0 @@
-export { VitestTestRunner } from './runtime/runners/test'
-export { NodeBenchmarkRunner } from './runtime/runners/benchmark'
