--- packages/vitest/src/snapshot.ts
@@ -1,2 +1,0 @@
-export type { SnapshotEnvironment } from '@vitest/snapshot/environment'
-export { VitestNodeSnapshotEnvironment as VitestSnapshotEnvironment } from './integrations/snapshot/environments/node'
