--- packages/vitest/src/workers.ts
@@ -1,10 +1,0 @@
-export {
-  createForksRpcOptions,
-  createThreadsRpcOptions,
-  unwrapSerializableConfig,
-} from './runtime/workers/utils'
-export { provideWorkerState } from './runtime/utils'
-export { run as runVitestWorker, collect as collectVitestWorkerTests } from './runtime/worker'
-export { runVmTests } from './runtime/workers/vm'
-export { runBaseTests } from './runtime/workers/base'
-export type { WorkerRpcOptions, VitestWorker } from './runtime/workers/types'
