+++ tsconfig.base.json
@@ -23,10 +23,10 @@
       "vitest/globals": ["./packages/vitest/globals.d.ts"],
       "vitest/node": ["./packages/vitest/src/public/node.ts"],
       "vitest/execute": ["./packages/vitest/src/public/execute.ts"],
-      "vitest/config": ["./packages/vitest/src/config.ts"],
-      "vitest/coverage": ["./packages/vitest/src/coverage.ts"],
-      "vitest/browser": ["./packages/vitest/src/browser.ts"],
-      "vitest/runners": ["./packages/vitest/src/runners.ts"],
+      "vitest/config": ["./packages/vitest/src/public/config.ts"],
+      "vitest/coverage": ["./packages/vitest/src/public/coverage.ts"],
+      "vitest/browser": ["./packages/vitest/src/public/browser.ts"],
+      "vitest/runners": ["./packages/vitest/src/public/runners.ts"],
       "vite-node": ["./packages/vite-node/src/index.ts"],
       "vite-node/client": ["./packages/vite-node/src/client.ts"],
       "vite-node/server": ["./packages/vite-node/src/server.ts"],
